+++ shared/src/commonMain/kotlin/org/hyperskill/app/next_learning_activity_widget/presentation/NextLearningActivityWidgetFeature.kt
@@ -2,8 +2,8 @@ package org.hyperskill.app.next_learning_activity_widget.presentation
 
 import org.hyperskill.app.analytic.domain.model.AnalyticEvent
 import org.hyperskill.app.learning_activities.domain.model.LearningActivity
+import org.hyperskill.app.learning_activities.presentation.model.LearningActivityTargetViewAction
 import org.hyperskill.app.study_plan.domain.model.StudyPlan
-import org.hyperskill.app.study_plan.widget.presentation.model.LearningActivityTargetViewAction
 
 object NextLearningActivityWidgetFeature {
     data class State(
